speech-recognition/requirements.txt
SpeechRecognition>=3.10.0
requests>=2.31.0
groq>=0.3.0
python-dotenv>=1.0.0
pyaudio>=0.2.13